WIP Use HidSharp in Wiimote

+++ WiimoteCS/WiimoteLib2/Wiimote.cs
@@ -1,4 +1,5 @@
 ﻿using HidSharp;
+using WiimoteLib;
 
 namespace WiimoteLib2;
 
@@ -51,7 +52,7 @@ public sealed class Wiimote : IDisposable, IAsyncDisposable
 
     public async Task SetReportingModeAsync(ReportingMode reportingMode, bool continuous, CancellationToken ct)
     {
-        await HidStream.SetReportingModeAsync(reportingMode, continuous, ct);
+        await HidStream.SetReportingModeAsync((byte)reportingMode, continuous, ct);
     }
 
     public async Task ContinuouslyReceiveReportsAsync(CancellationToken ct)
